Processing feedback: changed annotation name to View, added javadoc
