+++ test2/faces40/javaPage/src/main/java/com/sun/faces/test/faces40/javapages/Hello.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.util.List;
 
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.faces.annotation.FacesViewId;
+import jakarta.faces.annotation.View;
 import jakarta.faces.component.UIComponent;
 import jakarta.faces.component.UIOutput;
 import jakarta.faces.component.html.HtmlBody;
@@ -16,7 +16,7 @@ import jakarta.faces.component.html.HtmlOutputText;
 import jakarta.faces.context.FacesContext;
 import jakarta.faces.view.facelets.Facelet;
 
-@FacesViewId("/hello.xhtml")
+@View("/hello.xhtml")
 @ApplicationScoped
 public class Hello extends Facelet {
